fix type int to int64

+++ src/Drone/Types/Cron.hs
@@ -4,21 +4,22 @@
 module Drone.Types.Cron where
 
 import           Data.Extensible
+import           Data.Int        (Int64)
 import           Data.Text       (Text)
 
 type Cron = Record
-   '[ "id"       >: Int
-    , "repo_id"  >: Int
+   '[ "id"       >: Int64
+    , "repo_id"  >: Int64
     , "name"     >: Text
     , "expr"     >: Text
-    , "next"     >: Int
-    , "prev"     >: Int
+    , "next"     >: Int64
+    , "prev"     >: Int64
     , "event"    >: Text
     , "branch"   >: Text
     -- , "target"   >: Text
     , "disabled" >: Bool
-    , "created"  >: Int
-    , "updated"  >: Int
+    , "created"  >: Int64
+    , "updated"  >: Int64
     ]
 
 type CronPatch = Nullable (Field Identity) :* CronPatchFields
